Fixing bugs

+++ PhoneBook/phonebook-api/src/main/java/com/livebox/phonebookapi/model/CreateOrUpdateContactRequest.java
@@ -11,5 +11,5 @@ public class CreateOrUpdateContactRequest {
 
     String lastName;
 
-    Integer phoneNumber;
+    Long phoneNumber;
 }
